+++ Compiler/Controllers/ExternalSymbolControllers.Tests/ExternalVariableSymbolControllerTest.cs
@@ -2,7 +2,6 @@ using System.IO;
 
 using KSPCompiler.ExternalSymbolRepository.Tsv.Variables;
 using KSPCompiler.ExternalSymbolRepository.Yaml.Variables;
-using KSPCompiler.ExternalSymbolRepository.Yaml.Variables.v1;
 using KSPCompiler.Infrastructures.Commons.LocalStorages;
 using KSPCompiler.Interactor.Symbols;
