+++ Compiler/Infrastructures/ExternalSymbolRepository.Yaml.Tests/VariableTableYamlLoaderTest.cs
@@ -4,8 +4,7 @@ using System.Threading.Tasks;
 
 using KSPCompiler.ExternalSymbol.Commons;
 using KSPCompiler.ExternalSymbolRepository.Yaml.Variables;
-using KSPCompiler.ExternalSymbolRepository.Yaml.Variables.v1;
-using KSPCompiler.ExternalSymbolRepository.Yaml.Variables.v1.Model;
+using KSPCompiler.ExternalSymbolRepository.Yaml.Variables.Model;
 using KSPCompiler.Infrastructures.Commons.LocalStorages;
 
 using NUnit.Framework;
