+++ Compiler/Infrastructures/ExternalSymbolRepository.Yaml/Commands/Translators/ToYamlTranslator.cs
@@ -1,9 +1,9 @@
 using KSPCompiler.Commons;
 using KSPCompiler.Domain.Symbols;
 using KSPCompiler.Domain.Symbols.MetaData;
-using KSPCompiler.ExternalSymbolRepository.Yaml.Commands.v1.Model;
+using KSPCompiler.ExternalSymbolRepository.Yaml.Commands.Model;
 
-namespace KSPCompiler.ExternalSymbolRepository.Yaml.Commands.v1.Translators;
+namespace KSPCompiler.ExternalSymbolRepository.Yaml.Commands.Translators;
 
 internal class ToYamlTranslator : IDataTranslator<ISymbolTable<CommandSymbol>, RootObject>
 {
