+++ Compiler/Infrastructures/ExternalSymbolRepository.Yaml/Variables/Translators/FromYamlTranslator.cs
@@ -2,9 +2,9 @@ using KSPCompiler.Commons;
 using KSPCompiler.Domain.Symbols;
 using KSPCompiler.Domain.Symbols.MetaData;
 using KSPCompiler.ExternalSymbol.Commons;
-using KSPCompiler.ExternalSymbolRepository.Yaml.Variables.v1.Model;
+using KSPCompiler.ExternalSymbolRepository.Yaml.Variables.Model;
 
-namespace KSPCompiler.ExternalSymbolRepository.Yaml.Variables.v1.Translators;
+namespace KSPCompiler.ExternalSymbolRepository.Yaml.Variables.Translators;
 
 internal class FromYamlTranslator : IDataTranslator<RootObject, ISymbolTable<VariableSymbol>>
 {
